+++ NEWS.md
@@ -1,6 +1,7 @@
 # tectonicr 0.3.10 _2024-07-13_
 
 * more statistical estimators
+* jittered circular dot plot
 
 # tectonicr 0.3.9 _2024-06-20_